+++ src/libs/state.js
@@ -53,6 +53,8 @@ export const snapshotsListStore = Utils.atom()
 
 export const snapshotStore = Utils.atom()
 
+export const dataCatalogStore = Utils.atom()
+
 /*
  * Modifies ajax responses for testing purposes.
  * Can be set to an array of objects of the form { fn, filter }.
